requirement.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.5
aiosignal==1.4.0
alembic==1.18.4
amqp==5.3.1
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.13.0
argon2-cffi==25.1.0
argon2-cffi-bindings==25.1.0
asyncpg==0.31.0
attrs==26.1.0
bcrypt==5.0.0
billiard==4.2.4
black==26.3.1
celery==5.6.3
certifi==2026.2.25
cffi==2.0.0
click==8.3.2
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
DateTime==6.0
dnspython==2.8.0
email-validator==2.3.0
fastapi==0.135.3
fastapi-cache2==0.2.2
fastapi-cli==0.0.24
fastapi-cloud-cli==0.16.1
fastar==0.10.0
frozenlist==1.8.0
Funk==0.5.0
greenlet==3.4.0
h11==0.16.0
httpcore==1.0.9
httptools==0.7.1
httpx==0.28.1
idna==3.11
Jinja2==3.1.6
kombu==5.6.2
Mako==1.3.10
markdown-it-py==4.0.0
MarkupSafe==3.0.3
mdurl==0.1.2
multidict==6.7.1
mypy_extensions==1.1.0
packaging==26.0
passlib==1.7.4
pathspec==1.0.4
pendulum==3.2.0
pillow==12.2.0
platformdirs==4.9.6
precisely==0.1.9
prompt_toolkit==3.0.52
propcache==0.4.1
pwdlib==0.3.0
pycparser==3.0
pydantic==2.12.5
pydantic-extra-types==2.11.2
pydantic-settings==2.13.1
pydantic_core==2.41.5
Pygments==2.20.0
PyJWT==2.12.1
python-dateutil==2.9.0.post0
python-dotenv==1.2.2
python-multipart==0.0.24
pytokens==0.4.1
pytz==2026.1.post1
PyYAML==6.0.3
redis==7.4.0
rich==14.3.3
rich-toolkit==0.19.7
rignore==0.7.6
sentry-sdk==2.57.0
shellingham==1.5.4
six==1.17.0
SQLAlchemy==2.0.49
starlette==1.0.0
typer==0.24.1
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2026.1
tzlocal==5.3.1
unicorn==2.1.4
urllib3==2.6.3
uvicorn==0.44.0
uvloop==0.22.1
vine==5.1.0
watchfiles==1.1.1
wcwidth==0.6.0
websockets==16.0
yarl==1.23.0
zope.interface==8.3
